Code/openstreetmap_python/aptrequirements.txt
jquery
leaflet
python3-flask
python3-flask-sqlalchemy
